+++ src/lib/server/container/index.ts
@@ -110,6 +110,9 @@ function createPrismaContainer(): Container {
   const {
     PrismaCalendarioRepository,
   } = require("@/lib/server/db/implementations/prisma/prisma-calendario-repository");
+  const {
+    PrismaDisciplinaSemestreRepository,
+  } = require("@/lib/server/db/implementations/prisma/prisma-disciplina-semestre-repository");
 
   return {
     usuariosRepository: new PrismaUsuariosRepository(),
@@ -125,6 +128,7 @@ function createPrismaContainer(): Container {
     cargosRepository: new PrismaCargosRepository(),
     curriculosRepository: new PrismaCurriculosRepository(),
     calendarioRepository: new PrismaCalendarioRepository(),
+    disciplinaSemestreRepository: new PrismaDisciplinaSemestreRepository(),
     emailService: getEmailService(),
     blobStorage: getBlobStorage(),
   };